requirements.txt
Django>=4.0
requests
python-dotenv
discord.py
python-telegram-bot
railway
gunicorn